requirements.txt
matplotlib==3.9.1
numpy==2.2.6
pandas==2.2.3
Pillow==11.2.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
torch==2.6.0
torchvision==0.21.0
